Refactor according to PR discussion
